Refactored the servlets, inherits BaseServlet
