+++ src/main/java/com/bookstore/controller/admin/user/DeleteUserServlet.java
@@ -5,8 +5,7 @@ import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
-import com.bookstore.controller.admin.BaseServlet;
+import com.bookstore.controller.BaseServlet;
 import com.bookstore.service.UserServices;
 
 @WebServlet("/admin/delete_user")
