+++ src/main/java/com/bookstore/controller/frontend/HomeServlet.java
@@ -1,16 +1,20 @@
 package com.bookstore.controller.frontend;
 
 import java.io.IOException;
+import java.util.List;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.bookstore.controller.BaseServlet;
+import com.bookstore.dao.CategoryDAO;
+import com.bookstore.entity.Category;
+
 @WebServlet("")
-public class HomeServlet extends HttpServlet {
+public class HomeServlet extends BaseServlet {
 	private static final long serialVersionUID = 1L;
 
     public HomeServlet() {
@@ -18,6 +22,11 @@ public class HomeServlet extends HttpServlet {
     }
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		
+		CategoryDAO categoryDAO = new CategoryDAO(entityManager);
+		List<Category> listCategory = categoryDAO.listAll();
+		request.setAttribute("listCategory", listCategory);
+		
 		String page = "frontend/index.jsp";
 		RequestDispatcher requestDispatcher = request.getRequestDispatcher(page);
 		requestDispatcher.forward(request, response);
