+++ src/main/webapp/admin/header.jsp
@@ -1,3 +1,4 @@
+
 <div align="center">
 	<div>
 		<img src="../images/BookstoreAdminLogo.png" />
