+++ src/main/webapp/frontend/footer.jsp
@@ -3,5 +3,5 @@
 		<a href="about">About Us</a> |
 	<a href="contact">Contact Us</a> |
 	<a href="policy">Privacy Policy</a> |
-	<a href="shipping">Shipping & Delivery</a>
+	<a href="shipping">Shipping and Delivery</a>
 </div>
