+++ src/main/webapp/frontend/header.jsp
@@ -1,6 +1,7 @@
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <div align="center">
 	<div>
-		<img src="images/BookstoreLogo.png" />
+		<img src="images/BookstoreLogo.png"/>
 	</div>
 	
 	<div>
@@ -11,5 +12,16 @@
 		<a href="Login">Sign In</a>  |
 		<a href="register">Register</a> |
 		<a href="view_cart">Cart</a>
+	</div>
+	<div>&nbsp;</div>
+	<div align = "center">
+		<c:forEach var = "category" items="${listCategory}" varStatus="status">
+			<a href="view_category?id=${category.categoryId}">
+				<font size="+1"><b><c:out value= "${category.name}"/> </b></font>
+			</a>
+			<c:if test="${not status.last}">
+				&nbsp; | &nbsp;
+			</c:if>
+		</c:forEach>
 	</div>
 </div>
